Animation shader fixes

+++ GameClient/src/main/resources/assets/game/shaders/default.vert
@@ -3,7 +3,7 @@
 layout (location=0) in vec3 position;
 layout (location=1) in vec3 normal;
 layout (location=2) in vec2 textureCoord;
-layout (location=3) in int boneIndex;
+layout (location=3) in float boneIndex;
 
 const int MAX_BONES = 128;
 
@@ -16,8 +16,8 @@ uniform mat4 boneTransformations[MAX_BONES];
 
 void main()
 {
-    gl_Position = projectionMatrix * modelMatrix * boneTransformations[boneIndex] * vec4(position, 1.0);
-    vec4 animatedNormal = boneTransformations[boneIndex] * vec4(normal, 1.0);
+    gl_Position = projectionMatrix * modelMatrix * boneTransformations[int(boneIndex)] * vec4(position, 1.0);
+    vec4 animatedNormal = boneTransformations[int(boneIndex)] * vec4(normal, 1.0);
     outNormal = normalize(animatedNormal.xyz);
     outTextureCoord = textureCoord;
 }
